TDD iteration 7 - Refactoring Split test methods to 3 classes: without clients, with clients, with saving accounts
